requirements.txt
PySide6>=6.7.0
pika>=1.3.2
PyYAML>=6.0.1
